requirements.txt
python-telegram-bot==20.3
python-binance==1.0.29
python-dotenv==1.1.1
aiohttp==3.8.4
